refactor(cms): streamline custom field handling in post routes
- Introduced new functions for building and writing custom field values in post routes, enhancing the management of custom fields.
- Updated imports to include necessary types and utilities for improved functionality.
- Removed obsolete custom field functions from the custom-fields module to maintain clarity and reduce redundancy.
